remove some non-needed code

+++ xa.src/xa.cc
@@ -581,7 +581,7 @@ cl_xa::disass(t_addr addr, const char *sep)
     case BIT_REL8 :
       sprintf(parm_str, "%s,0x%04lx",
 	      get_bit_name((code&0x0003)<<8) + get_mem(MEM_ROM_ID, addr+2),
-	      ((signed char)get_mem(MEM_ROM_ID, addr+3)*2+addr+len)&0xfffe);
+	      (long)(((signed char)get_mem(MEM_ROM_ID, addr+3)*2+addr+len)&0xfffe));
     break;
     case DATA4:
       sprintf(parm_str, "#0x%02x", code&0x0f);
@@ -596,13 +596,13 @@ cl_xa::disass(t_addr addr, const char *sep)
     case REG_REL8 :
       sprintf(parm_str, "%s,0x%04lx",
 	      reg_strs[(code>>4) & 0xf],
-	      ((signed char)get_mem(MEM_ROM_ID, addr+2)*2+addr+len)&0xfffe);
+	      (long)(((signed char)get_mem(MEM_ROM_ID, addr+2)*2+addr+len)&0xfffe));
     break;
     case DIRECT_REL8 :
       sprintf(parm_str, "%s,0x%04lx",
 	      get_dir_name(((code&0x07)<<8) +
 			   get_mem(MEM_ROM_ID, addr+2)),
-	      ((signed char)get_mem(MEM_ROM_ID, addr+2)*2+addr+len)&0xfffe);
+	      (long)(((signed char)get_mem(MEM_ROM_ID, addr+2)*2+addr+len)&0xfffe));
     break;
     case REG_USP:
       sprintf(parm_str, "REG_USP");
@@ -612,11 +612,11 @@ cl_xa::disass(t_addr addr, const char *sep)
     break;
     case REL8 :
       sprintf(parm_str, "0x%04lx",
-	      ((signed char)get_mem(MEM_ROM_ID, addr+1)*2+addr+len)&0xfffe);
+	      (long)(((signed char)get_mem(MEM_ROM_ID, addr+1)*2+addr+len)&0xfffe));
     break;
     case REL16 :
       sprintf(parm_str, "0x%04lx",
-	      ((signed short)((get_mem(MEM_ROM_ID, addr+1)<<8) + get_mem(MEM_ROM_ID, addr+2))*2+addr+len)&0xfffe);
+	      (long)(((signed short)((get_mem(MEM_ROM_ID, addr+1)<<8) + get_mem(MEM_ROM_ID, addr+2))*2+addr+len)&0xfffe));
     break;
 
     case RLIST : {
